service.CashFileSequenceService renamed and moved to new package
model.CashFileSequence
